+++ src/components/Cards/CardInfo/cardInfo.css
